Cambios en InscripcionData e Inscripcion

+++ universidadejemplo/universidadEjemplo/src/universidadejemplo/entidades/Inscripcion.java
@@ -5,25 +5,25 @@ public class Inscripcion {
     int idInscripcion;
     Materia materia;
     Alumno alumno;
-    int nota;
+    double nota;
 
     public Inscripcion() {
     }
 
-    public Inscripcion(int idInscripcion, Materia materia, Alumno alumno, int nota) {
+    public Inscripcion(int idInscripcion, Materia materia, Alumno alumno, double nota) {
         this.idInscripcion = idInscripcion;
         this.materia = materia;
         this.alumno = alumno;
         this.nota = nota;
     }
 
-    public Inscripcion(Materia materia, Alumno alumno, int nota) {
+    public Inscripcion(Materia materia, Alumno alumno, double nota) {
         this.materia = materia;
         this.alumno = alumno;
         this.nota = nota;
     }
 
-    public Inscripcion(int nota) {
+    public Inscripcion(double nota) {
         this.nota = nota;
     }
 
@@ -51,11 +51,11 @@ public class Inscripcion {
         this.alumno = alumno;
     }
 
-    public int getNota() {
+    public double getNota() {
         return nota;
     }
 
-    public void setNota(int nota) {
+    public void setNota(double nota) {
         this.nota = nota;
     }
